copy_system: Add copy system script
Copy system is a script to automatically copy the root filesystem to another
disk, formatting it to whatever filesystem is wished (ext4 being the current
default).

Also only one partition is created on the target disk and syslinux is assumed
to be the bootloader.

The script is intented to copy the runnig rootfs from a USB stick to another
medium like a second USB stick or a mSATA disk.

Signed-off-by: Johannes Thumshirn <johannes.thumshirn@men.de>